Revert "Quitar código duplicado"
