+++ src/model/data_structures/TablaHashSeparteChaining.java
@@ -2,8 +2,16 @@ package model.data_structures;
 
 import java.text.DecimalFormat;
 
-public class TablaHashSeparteChaining <K extends Comparable<K>, V extends Comparable <V>> extends TablaHash<K, V, ILista<NodoTS<K,V>>>
+import sun.security.util.Debug;
+
+public class TablaHashSeparteChaining <K extends Comparable<K>, V extends Comparable <V>> implements ITablaSimbolos<K, V >
 {
+	
+	private ILista<ILista<NodoTS<K,V>>> listaNodos;
+	private int tamanoAct;
+	private int tamanoTabla;
+	private double minicial;
+	private double cantidadRehash;
 	
 	public TablaHashSeparteChaining(int tamInicial)
 	{
@@ -244,6 +252,84 @@ public class TablaHashSeparteChaining <K extends Comparable<K>, V extends Compar
 		return Math.abs(key.hashCode()% tamanoTabla) +1;
 	}
 	
+	public void rehash()
+	{
+		try
+		{
+			ILista<NodoTS<K,V>> nodos= darListaNodos();
+			
+			tamanoAct=0;
+			tamanoTabla*=2;
+			int m = nextPrime(tamanoTabla);
+			tamanoTabla=m;
+			listaNodos=new ArregloDinamico<>(tamanoTabla);
+			
+			for(int i=1; i<=tamanoTabla; i++)
+			{
+				listaNodos.insertElement(null, size()+1);
+			}
+			
+			NodoTS<K,V> actual= null;
+			for(int i=1; i<= nodos.size();i++)
+			{
+				actual=nodos.getElement(i);
+				put(actual.getKey(), actual.getValue());
+			}
+		}
+		catch (NullException| VacioException| PosException e)
+		{
+			e.printStackTrace();
+		}
+		
+		cantidadRehash++;
+		
+	}
+	
+	static boolean isPrime(int n)
+    {
+
+        if (n <= 1) return false;
+
+        if (n > 1 && n <= 3) return true;
+
+
+        if (n % 2 == 0 || n % 3 == 0) return false;
+
+        for (int i = 5; i * i <= n; i = i + 6)
+
+            if (n % i == 0 || n % (i + 2) == 0)
+
+            return false;
+
+        return true;
+    }
+
+    static int nextPrime(int N)
+
+    {
+        if (N <= 1)
+
+            return 2;
+
+        int prime = N;
+
+        boolean found = false;
+
+
+        while (!found)
+
+        {
+            prime++;
+
+            if (isPrime(prime))
+
+                found = true;
+
+        }
+        return prime;
+
+    }
+    
     public String toString()
     {
     	String retorno="";
